backend/pyproject.toml
# NEWS CHUNK 3 — Backend Routing & Test Endpoint
# Author: GPT-5 Codecs (acting as a 30–40 year experienced software engineer)
# Behavior: Full write access. Create files, run checks, save results.

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fake-news-prediction-backend"
version = "0.0.1"
description = "FastAPI backend scaffold for the fake news prediction project."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0,<1.0.0",
  "uvicorn[standard]>=0.29.0,<1.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.2.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "respx>=0.20.2,<1.0.0",
  "redis>=5.0.0,<6.0.0"
]
